pyproject.toml
[project]
name = "satisfaction-dashboard"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "bcrypt>=5.0.0",
    "google-auth>=2.28.0",
    "gspread>=6.2.1",
    "pandas>=2.3.3",
    "plotly>=6.6.0",
    "pyyaml>=6.0.3",
    "streamlit>=1.55.0",
    "streamlit-authenticator>=0.4.2",
]
